remove Endec.int & Endec.double, replace with proper sized variants

+++ endec_tests/test/flat_struct_field_test.dart
@@ -6,13 +6,13 @@ void main() {
   test('encode child class to json', () {
     final fieldsEndec = structEndec<(String, int)>().with2Fields(
       Endec.string.fieldOf("a_field", (tuple) => tuple.$1),
-      Endec.int.fieldOf("another_field", (tuple) => tuple.$2),
+      Endec.i64.fieldOf("another_field", (tuple) => tuple.$2),
       (p0, p1) => (p0, p1),
     );
 
     final childClassEndec = structEndec<ChildClass>().with2Fields(
       fieldsEndec.flatFieldOf((struct) => (struct.aField, struct.anotherField)),
-      Endec.double.listOf().fieldOf("third_field", (struct) => struct.thirdField),
+      Endec.f64.listOf().fieldOf("third_field", (struct) => struct.thirdField),
       (p0, p1) => ChildClass(p0.$1, p0.$2, p1),
     );
 
@@ -29,13 +29,13 @@ void main() {
   test('encode grandchild class to json', () {
     final fieldsEndec = structEndec<(String, int)>().with2Fields(
       Endec.string.fieldOf("a_field", (tuple) => tuple.$1),
-      Endec.int.fieldOf("another_field", (tuple) => tuple.$2),
+      Endec.i64.fieldOf("another_field", (tuple) => tuple.$2),
       (p0, p1) => (p0, p1),
     );
 
     final childClassEndec = structEndec<ChildClass>().with2Fields(
       fieldsEndec.flatFieldOf((struct) => (struct.aField, struct.anotherField)),
-      Endec.double.listOf().fieldOf("third_field", (struct) => struct.thirdField),
+      Endec.f64.listOf().fieldOf("third_field", (struct) => struct.thirdField),
       (p0, p1) => ChildClass(p0.$1, p0.$2, p1),
     );
